dev/performance/requirements.txt
Faker
mysql-connector-python==8.0.31